[New Physical IO] Sequential Prefetching Implementation (#289)
## Description of change
This PR adds a new method optimizeReads to the RangeOptimiser class to
improve read performance by intelligently grouping and splitting block
indexes. The implementation reduces the complexity in DataBlockManager
and makes the optimization logic more testable.

Changes are: 
- Adds readAheadBytes logic
- Adds sequential prefetching logic
- Groups sequential block indexes together
- Splits large sequential groups into smaller chunks based on
configuration parameters
- Refactored DataBlockManager to use the new method instead of
implementing the logic itself
- Added comprehensive unit tests for the new method

Out of Scope
- Range coalescing will be implemented in a separate PR


#### Relevant issues
PR History:

https://github.com/awslabs/analytics-accelerator-s3/pull/286
https://github.com/awslabs/analytics-accelerator-s3/pull/287
https://github.com/awslabs/analytics-accelerator-s3/pull/288

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).