Add request coalescing and relevant configuration (#356)
## Description of change
With this change, AAL can coalesce requests in a close proximity to each
other.

This enables to take spatial locality
into account and reduce request count. In my testing I observed that
this is particularly important for vectored reads
reading vectors near each other. 

There are two configurations to manage this feature. First one is to
enable disable request coalescing:
`request.coalesce` which is by default set to `true`.

Second one is `request.coalesce.tolerance`, which is the number of bytes
to tolerate (read even if not immediately needed) when merging ranges.

Two ranges will be merged if end of the first range is less than
`tolerate` bytes away from the second one.

#### Relevant issues
N/A

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No. It changes the behaviour of vectored reads and merges smaller
requests to bigger requests.
These bigger requests might be split to smaller requests later to make
sure all requests are around `target.request.size`

#### Does this contribution introduce any new public APIs or behaviors?
<!-- Please describe them and explain what scenarios they target.  -->

#### How was the contribution tested?
Added new unit tests. Extended existing IntegrationTests to run for both
coalescing enabled and disabled.

#### Does this contribution need a changelog entry?
- Yes, i shall be updating the README after we finalise the behaviour
part of this PR.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).